dod.tideGauge() handles new CHS data (WIP)
This is a WIP because I am not settled on the column names.  The NOAA
names, which is what we have now, are sort of stupid.  Why is one "Water
Level" (singular) while the other is "Predictions" (plural)?  And why is
time given as "Date Time"?  The oce scheme would more sensibly be

* time
* height

(I don't like "level" because the water is not level)